refactor(apps/resume): 컴포넌트 명칭 수정
